+++ app/src/main/java/com/example/workschedule/data/database/train/TrainDao.kt
@@ -1,4 +1,4 @@
-package com.example.workschedule.data.db.train
+package com.example.workschedule.data.database.train
 
 
 import androidx.room.*
